Refactored all the methods to use other classes functionality into FinancialPlannerSystemTest
